Fixed bug in training plan workout text related to debounce logic and overwriting newer user input
